First working version of plugin (ref #311)
