+++ .coveragerc
@@ -1,4 +1,5 @@
 [run]
 omit = kinto/tests/*
+       kinto/plugins/*/test_*
        kinto/__main__.py
